Empty labels count as not querying on label.
In version 0.1.3 a changed was introduced that made queries with an
empty label return no results and not all results. This was somewhat of
an attempt to placate a dojo datastore further down the line. Tying
skosprovider to a specific client implementation seems like a bad idea.
So I've gone to what seems to be the most sensible solution: calling
find with an empty label is the same as calling it without a label and
so returns every concept. Other filterings still take place and can
modify the result.